+++ examples/async.rs
@@ -22,39 +22,41 @@ pub struct Context {
 
 #[async_trait]
 impl StateMachineContext for Context {
-    fn guard3(&mut self) -> Result<bool, ()> {
+    fn guard3(&self) -> Result<bool, ()> {
         println!("`guard3` called from async context");
         Ok(true)
     }
 
-    async fn guard2(&mut self) -> Result<bool, ()> {
+    async fn guard2(&self) -> Result<bool, ()> {
         println!("`guard2` called from async context");
         let mut lock = self.lock.write().await;
         *lock = false;
         Ok(true)
     }
 
-    fn guard1(&mut self) -> Result<bool, ()> {
+    fn guard1(&self) -> Result<bool, ()> {
         println!("`guard1` called from sync context");
         Ok(true)
     }
 
-    async fn action2(&mut self) -> () {
+    async fn action2(&mut self) -> Result<(), ()> {
         println!("`action2` called from async context");
         if !*self.lock.read().await {
             self.done = true;
         }
+        Ok(())
     }
 
-    async fn action1(&mut self) -> () {
+    async fn action1(&mut self) -> Result<(), ()> {
         println!("`action1` called from async context");
         let mut lock = self.lock.write().await;
         *lock = true;
+        Ok(())
     }
 
-    fn action3(&mut self) -> bool {
+    fn action3(&mut self) -> Result<bool, ()> {
         println!("`action3` called from sync context, done = `{}`", self.done);
-        self.done
+        Ok(self.done)
     }
 }