+++ examples/guard_action_syntax.rs
@@ -27,22 +27,22 @@ pub struct Context;
 
 impl StateMachineContext for Context {
     // Guard1 has access to the data from Event1
-    fn guard1(&mut self, _event_data: &MyEventData) -> Result<bool, ()> {
+    fn guard1(&self, _event_data: &MyEventData) -> Result<bool, ()> {
         todo!()
     }
 
     // Action1 has access to the data from Event1, and need to return the state data for State2
-    fn action1(&mut self, _event_data: MyEventData) -> MyStateData {
+    fn action1(&mut self, _event_data: MyEventData) -> Result<MyStateData, ()> {
         todo!()
     }
 
     // Guard2 has access to the data from State2
-    fn guard2(&mut self, _state_data: &MyStateData) -> Result<bool, ()> {
+    fn guard2(&self, _state_data: &MyStateData) -> Result<bool, ()> {
         todo!()
     }
 
     // Action2 has access to the data from State2
-    fn action2(&mut self, _state_data: &MyStateData) {
+    fn action2(&mut self, _state_data: &MyStateData) -> Result<(), ()> {
         todo!()
     }
 }