+++ examples/starting_state_with_data.rs
@@ -22,8 +22,8 @@ statemachine! {
 pub struct Context;
 
 impl StateMachineContext for Context {
-    fn action(&mut self) -> MyStateData {
-        MyStateData(42)
+    fn action(&mut self) -> Result<MyStateData, ()> {
+        Ok(MyStateData(42))
     }
 }