+++ tests/test.rs
@@ -48,27 +48,33 @@ fn multiple_lifetimes() {
     struct Context;
 
     impl StateMachineContext for Context {
-        fn guard1(&mut self, _event_data: &X) -> Result<bool, ()> {
+        fn guard1(&self, _event_data: &X) -> Result<bool, ()> {
             Ok(true)
         }
 
-        fn guard2(&mut self, _state_data: &X, _event_data: &Y) -> Result<bool, ()> {
+        fn guard2(&self, _state_data: &X, _event_data: &Y) -> Result<bool, ()> {
             Ok(true)
         }
 
-        fn guard3(&mut self, _event_data: &Z) -> Result<bool, ()> {
+        fn guard3(&self, _event_data: &Z) -> Result<bool, ()> {
             Ok(true)
         }
 
-        fn action1<'a>(&mut self, event_data: &'a X) -> &'a X {
-            event_data
+        fn action1<'a>(&mut self, event_data: &'a X) -> Result<&'a X, ()> {
+            Ok(event_data)
         }
 
-        fn action2<'a, 'b>(&mut self, state_data: &'a X, event_data: &'b Y) -> (&'a X, &'b Y) {
-            (state_data, event_data)
+        fn action2<'a, 'b>(
+            &mut self,
+            state_data: &'a X,
+            event_data: &'b Y,
+        ) -> Result<(&'a X, &'b Y), ()> {
+            Ok((state_data, event_data))
         }
 
-        fn action3(&mut self, _event_data: &Z) {}
+        fn action3(&mut self, _event_data: &Z) -> Result<(), ()> {
+            Ok(())
+        }
     }
 
     #[allow(dead_code)]
@@ -140,12 +146,12 @@ fn async_guards_and_actions() {
         struct Context;
         #[smlang::async_trait]
         impl StateMachineContext for Context {
-            async fn guard1(&mut self) -> Result<bool, ()> {
+            async fn guard1(&self) -> Result<bool, ()> {
                 Ok(true)
             }
 
-            async fn action1(&mut self) -> () {
-                ()
+            async fn action1(&mut self) -> Result<(), ()> {
+                Ok(())
             }
         }
 
@@ -180,17 +186,19 @@ fn guard_expressions() {
         attempts: u32,
     }
     impl StateMachineContext for Context {
-        fn valid_entry(&mut self, e: &Entry) -> Result<bool, ()> {
+        fn valid_entry(&self, e: &Entry) -> Result<bool, ()> {
             Ok(e.0 == self.password)
         }
-        fn too_many_attempts(&mut self, _e: &Entry) -> Result<bool, ()> {
+        fn too_many_attempts(&self, _e: &Entry) -> Result<bool, ()> {
             Ok(self.attempts >= 3)
         }
-        fn reset(&mut self) {
+        fn reset(&mut self) -> Result<(), ()> {
             self.attempts = 0;
+            Ok(())
         }
-        fn attempt(&mut self, _e: &Entry) {
+        fn attempt(&mut self, _e: &Entry) -> Result<(), ()> {
             self.attempts += 1;
+            Ok(())
         }
     }
 
@@ -252,12 +260,13 @@ fn guarded_transition_before_unguarded() {
         pub enabled: bool,
     }
     impl StateMachineContext for Context {
-        fn guard(&mut self) -> Result<bool, ()> {
+        fn guard(&self) -> Result<bool, ()> {
             Ok(self.enabled)
         }
 
-        fn disable(&mut self) {
+        fn disable(&mut self) -> Result<(), ()> {
             self.enabled = false;
+            Ok(())
         }
     }
     let mut sm = StateMachine::new(Context { enabled: true });
@@ -285,7 +294,7 @@ fn guard_errors() {
         pub guard_errors: bool,
     }
     impl StateMachineContext for Context {
-        fn guard(&mut self) -> Result<bool, ()> {
+        fn guard(&self) -> Result<bool, ()> {
             if self.guard_errors {
                 Err(())
             } else {